// api/src/modules/auth/dto/auth-response.dto.ts
// Dto for the response of the authentication process

import { Role } from '@prisma/client';

export class AuthResponseDto {
  accessToken: string;
  refreshToken: string;

  user: {
    id: string;
    email: string;
    firstName?: string | null;
    lastName?: string | null;
    role: Role;
  };
}